# swagger.json
{"swagger": "2.0", "info": {"title": "API for WINK Hakaton", "description": "При написании сценария важно придерживаться выбранного возрастного рейтинга, устанавливаемого на основании норм Возрастной классификации информационной продукции. Такой подход гарантирует попадание контента в оптимальные эфирные/кинопрокатные слоты и доступность на стриминговых платформах, что позволяет достичь целевой аудитории зрителей. Автоматическая проверка рейтинга по сценарию позволяет авторам и продюсерам заранее понимать, где текст может выйти за рамки выбранной категории - делая процесс производства быстрее, прозрачнее и экономичнее.", "contact": {"email": "work80@mail.ru"}, "serviceIdentity": false, "version": "v1"}, "host": "127.0.0.1:8000", "schemes": ["http"], "basePath": "/api/v1/wink", "consumes": ["application/json"], "produces": ["application/json"], "securityDefinitions": {"Bearer": {"type": "apiKey", "name": "Authorization", "in": "header"}}, "security": [{"Bearer": []}], "paths": {"/cinema/": {"get": {"operationId": "cinema_list", "description": "", "parameters": [], "responses": {"200": {"description": "", "schema": {"type": "array", "items": {"$ref": "#/definitions/IntermediateFiles"}}}}, "tags": ["cinema"]}, "post": {"operationId": "cinema_create", "description": "\nThat is get the data from request and save them.\nThen send a signal on AI for would be to begin load (parsing) the file\n", "parameters": [{"name": "data", "in": "body", "required": true, "schema": {"required": ["file_id"], "type": "object", "properties": {"file_id": {"title": "body", "description": "This is the id of the one file", "type": "integer"}}}}, {"name": "X-CSRFToken", "in": "header", "description": "The X-CSRFToken header. Token 'csrftoken you cant take from the cookie or to look the API's tage 'csrf'", "required": true, "type": "string"}], "responses": {"200": {"description": "If all nice means the response contain just the status code 200. ", "schema": {"type": "object", "properties": {"id": {"type": "integer", "example": 7}, "violations_quantity": {"description": "That is the static number and all the time this the variable will have the number 0", "type": "integer", "example": 0}, "refer": {"description": "That is the static string and all tne time this the variable will have the empty string", "type": "string", "example": ""}, "upload": {"description": "That is the index of file. It will be send to the parser.", "type": "integer", "example": 7}, "user": {"description": "That is the index of user. This user who sent the one file to the parser.", "type": "integer"}, "violations": {"description": "This is array from everything of violations - total list. Swagger key is 'violations' ", "type": "array", "items": {"type": "integer", "example": [1, 2, 3, 4, 5, 6]}}}}}, "400": {"description": "Request is invalid."}, "404": {"description": "File with id does not exist."}, "500": {"description": "Internal Server Error."}}, "tags": ["cinema"]}, "parameters": []}, "/cinema/{id}/": {"get": {"operationId": "cinema_read", "description": "", "parameters": [], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/IntermediateFiles"}}}, "tags": ["cinema"]}, "put": {"operationId": "cinema_update", "description": "", "parameters": [{"name": "data", "in": "body", "required": true, "schema": {"$ref": "#/definitions/IntermediateFiles"}}], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/IntermediateFiles"}}}, "tags": ["cinema"]}, "patch": {"operationId": "cinema_partial_update", "description": "", "parameters": [{"name": "data", "in": "body", "required": true, "schema": {"$ref": "#/definitions/IntermediateFiles"}}], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/IntermediateFiles"}}}, "tags": ["cinema"]}, "delete": {"operationId": "cinema_delete", "description": "", "parameters": [], "responses": {"204": {"description": ""}}, "tags": ["cinema"]}, "parameters": [{"name": "id", "in": "path", "description": "A unique integer value identifying this Intermediate_files.", "required": true, "type": "integer"}]}, "/csrftoken/": {"get": {"operationId": "csrftoken_list", "description": "csrf-token", "parameters": [], "responses": {"200": {"description": "CSRF токен", "headers": {"Set-Cookie": {"description": "CSRF token cookie", "type": "string", "example": "csrftoken=GiehvEXjNiYqp3bOVtA...; Path=/; Secure; HttpOnly; SameSite=Lax"}}}}, "tags": ["csrf"]}, "parameters": []}, "/download/": {"get": {"operationId": "download_list", "description": "", "parameters": [], "responses": {"200": {"description": "", "schema": {"type": "array", "items": {"$ref": "#/definitions/IntermediateFiles"}}}}, "tags": ["download"]}, "parameters": []}, "/download/{id}/": {"get": {"operationId": "download_read", "description": "тут от меня начинают скачивать\n:param request:\n:param args:\n:param kwargs:\n:return:", "parameters": [], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/IntermediateFiles"}}}, "tags": ["download"]}, "parameters": [{"name": "id", "in": "path", "description": "A unique integer value identifying this Intermediate_files.", "required": true, "type": "integer"}]}, "/files/": {"get": {"operationId": "files_list", "description": "", "parameters": [], "responses": {"200": {"description": "", "schema": {"type": "array", "items": {"$ref": "#/definitions/Files"}}}}, "tags": ["files"]}, "post": {"operationId": "files_create", "description": "This method record meta-data about files and the 'upload' field is do empty. It's for the main flow.\n            Then, It create the new thread and inside of thread it upload the file to the server by server's path 'media/upload/%Y/%m/%d/file_name.{pdf|docx}'\n            Then, column 'upload' is updating from the last row's empty in the database.\n            Note: Record of file to the server and send it on parsing - this is two separate processes!\n            That is just the process a saving.\n", "parameters": [{"name": "data", "in": "body", "required": true, "schema": {"required": ["upload"], "type": "object", "properties": {"upload": {"title": "formData", "description": "The upload file. 'Content-Type: multipart/form-data' is request.", "type": "file"}}}}, {"name": "X-CSRFToken", "in": "header", "description": "The X-CSRFToken header. Token 'csrftoken you cant take from the cookie or to look the API's tage 'csrf'", "required": true, "type": "string"}], "responses": {"200": {"description": "If all nice means the response contain just the status code 200. ", "schema": {"type": "object", "properties": {"id": {"type": "integer", "example": 12}}}}, "400": {"description": "{'errors': 'text of error'}"}}, "tags": ["files"]}, "parameters": []}, "/files/{id}/": {"get": {"operationId": "files_read", "description": "", "parameters": [], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/Files"}}}, "tags": ["files"]}, "put": {"operationId": "files_update", "description": "", "parameters": [{"name": "data", "in": "body", "required": true, "schema": {"$ref": "#/definitions/Files"}}], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/Files"}}}, "tags": ["files"]}, "patch": {"operationId": "files_partial_update", "description": "", "parameters": [{"name": "data", "in": "body", "required": true, "schema": {"$ref": "#/definitions/Files"}}], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/Files"}}}, "tags": ["files"]}, "delete": {"operationId": "files_delete", "description": "", "parameters": [], "responses": {"204": {"description": ""}}, "tags": ["files"]}, "parameters": [{"name": "id", "in": "path", "description": "A unique integer value identifying this Файл.", "required": true, "type": "integer"}]}, "/violations/": {"get": {"operationId": "violations_list", "description": "", "parameters": [], "responses": {"200": {"description": "", "schema": {"type": "array", "items": {"$ref": "#/definitions/BasisViolation"}}}}, "tags": ["violations"]}, "post": {"operationId": "violations_create", "description": "", "parameters": [{"name": "data", "in": "body", "required": true, "schema": {"$ref": "#/definitions/BasisViolation"}}], "responses": {"201": {"description": "", "schema": {"$ref": "#/definitions/BasisViolation"}}}, "tags": ["violations"]}, "parameters": []}, "/violations/{id}/": {"get": {"operationId": "violations_read", "description": "", "parameters": [], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/BasisViolation"}}}, "tags": ["violations"]}, "put": {"operationId": "violations_update", "description": "", "parameters": [{"name": "data", "in": "body", "required": true, "schema": {"$ref": "#/definitions/BasisViolation"}}], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/BasisViolation"}}}, "tags": ["violations"]}, "patch": {"operationId": "violations_partial_update", "description": "", "parameters": [{"name": "data", "in": "body", "required": true, "schema": {"$ref": "#/definitions/BasisViolation"}}], "responses": {"200": {"description": "", "schema": {"$ref": "#/definitions/BasisViolation"}}}, "tags": ["violations"]}, "delete": {"operationId": "violations_delete", "description": "", "parameters": [], "responses": {"204": {"description": ""}}, "tags": ["violations"]}, "parameters": [{"name": "id", "in": "path", "description": "A unique integer value identifying this Basis Violation.", "required": true, "type": "integer"}]}}, "definitions": {"IntermediateFiles": {"required": ["upload", "user"], "type": "object", "properties": {"id": {"title": "ID", "type": "integer", "readOnly": true}, "violations_quantity": {"title": "Violations Quantity", "type": "integer"}, "refer": {"title": "Reference", "description": "Reference link to the file - pdf, docx", "type": "string", "format": "uuid", "readOnly": true}, "upload": {"title": "Файл", "description": "Select existing the file id", "type": "integer"}, "user": {"title": "Пользователь", "description": "Select the user id", "type": "integer"}, "violations": {"description": "Violations - the views of violations", "type": "array", "items": {"description": "Violations - the views of violations", "type": "integer"}, "uniqueItems": true}}}, "Files": {"type": "object", "properties": {"id": {"title": "ID", "type": "integer", "readOnly": true}, "upload": {"title": "Файл", "description": "Upload the file - pdf, docx", "type": "string", "readOnly": true, "x-nullable": true, "format": "uri"}, "name": {"title": "Name", "type": "string", "maxLength": 150, "x-nullable": true}, "size": {"title": "Size", "type": "integer", "minimum": 0, "x-nullable": true}}}, "BasisViolation": {"required": ["violations", "violations_comment"], "type": "object", "properties": {"id": {"title": "ID", "type": "integer", "readOnly": true}, "violations": {"title": "Violations", "description": "Violations - the views of violations", "type": "string", "pattern": "^[A-Za-z- ,]+$", "maxLength": 50, "minLength": 1}, "violations_comment": {"title": "Violations_comments", "description": "Example - Нецензурная лексика - Мат, обсценная речь 16+ / 18+", "type": "string", "pattern": "^[A-Za-z- +/0-9]+$", "maxLength": 100, "minLength": 1}}}}}